requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.5
flake8==3.5.0
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
mccabe==0.6.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
